+++ cg/constants/observations.py
@@ -1,5 +1,13 @@
-"""Loqusdb related constants"""
+"""Loqusdb related constants."""
 
-from cgmodels.cg.constants import Pipeline
+from cgmodels.cg.constants import Pipeline, StrEnum
 
-LOQUSDB_SUPPORTED_PIPELINES = [Pipeline.MIP_DNA]
+
+LOQUSDB_SUPPORTED_PIPELINES = [Pipeline.MIP_DNA, Pipeline.BALSAMIC]
+
+
+class ObservationAnalysisTag(StrEnum):
+    PEDIGREE: str = "pedigree"
+    CHECK_PROFILE_GBCF: str = "snv-gbcf"
+    SNV_VARIANTS: str = "deepvariant"
+    SV_VARIANTS: str = "vcf-sv-research"
